chore: promote channel.subscription.end eventsub to v1
